Animation added

+++ lib/pages/home_page/widgets/temperature_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../../../utils/helpers.dart';
 
 class TemperatureWidget extends StatelessWidget {
-  final int temp;
+  final double temp;
   final String weatherCondition;
   final DateTime dateTime;
   const TemperatureWidget({
